+++ R/create_rmd.R
@@ -2,28 +2,32 @@
 #'
 #' @param pkgs.df Data.frame with package names, versions, and citation keys,
 #' as produced by [get_pkgs_info()].
+#' @param out.dir Directory to save the output document.
 #' @param bib.file Name of the file containing references in BibTeX format,
 #' as produced by [get_pkgs_info()].
 #' @param csl Optional. Citation style to format references.
 #' See <https://www.zotero.org/styles>.
 #' @param out.format Output format. One of: "html", docx" (Word), "pdf", "Rmd", or "md" (markdown).
 #' @param Rmd.file Name of the Rmarkdown file to be created.
-#' @param out.dir Directory to save the output document. Default is the working directory.
 #' @param include.RStudio Include RStudio?
-#' @param ... Further arguments for [render_citations()].
+#' @param ... Further arguments for [render_citations()]. Currently not used.
 #'
 #' @return An Rmarkdown file, if out.format = "Rmd", or a rendered document otherwise.
 #' @noRd
 #'
 
 create_rmd <- function(pkgs.df = NULL,
-                       bib.file = "grateful-refs.bib",
+                       out.dir = NULL,
+                       bib.file = "grateful-refs",
                        csl = NULL,
-                       Rmd.file = "grateful-report.Rmd",
+                       Rmd.file = "grateful-report",
                        out.format = c("html", "docx", "pdf", "Rmd", "md"),
-                       out.dir = getwd(),
                        include.RStudio = FALSE,
                        ...) {
+
+  if (is.null(out.dir)) {
+    stop("Please specify where to save the citation report, e.g. out.dir = getwd()")
+  }
 
   out.format <- match.arg(out.format)
 
@@ -32,14 +36,18 @@ create_rmd <- function(pkgs.df = NULL,
   # ensure CSL file is available, otherwise download
   if (!is.null(csl)) {
     if (!file.exists(file.path(out.dir, paste0(csl, ".csl")))) {
-      get_csl(csl)
+      get_csl(csl, out.dir = out.dir)
     }
   }
 
+
+  bib.file <- paste0(bib.file, ".bib")
+  Rmd.file <- paste0(Rmd.file, ".Rmd")
+
   if (out.dir != getwd()) {
-    bibfile <- file.path(out.dir, bib.file)
-    csl <- file.path(out.dir, csl)
-    filename <- file.path(out.dir, Rmd.file)
+    # bibfile <- file.path(out.dir, bib.file)
+    # csl <- file.path(out.dir, csl)
+    Rmd.file <- file.path(out.dir, Rmd.file)
   }
 
   yaml.header <- c(
@@ -69,11 +77,16 @@ create_rmd <- function(pkgs.df = NULL,
              con = Rmd.file)
 
   if (tolower(out.format) == "rmd") {
-    return(Rmd.file)
+    out.file <- Rmd.file
   } else {
-    render_citations(Rmd.file, out.format = out.format, out.dir = out.dir, ...)
+    out.file <- render_citations(Rmd.file,
+                                 out.format = out.format,
+                                 out.dir = out.dir,
+                                 ...)
     file.remove(Rmd.file)
   }
+
+  invisible(out.file)
 
 }