+++ tests/testthat/test-get_pkgs_info.R
@@ -1,5 +1,5 @@
 test_that("get_pkgs_info works", {
-  info <- get_pkgs_info(pkgs = c("renv", "remotes", "dplyr", "ggplot2"))
+  info <- get_pkgs_info(pkgs = c("renv", "remotes", "dplyr", "ggplot2"), out.dir = tempdir())
   expect_identical(info$pkg, c("remotes", "renv", "tidyverse"))
   expect_identical(info$citekeys, list("remotes", "renv", "tidyverse"))
 })